vat added

+++ schemas/order/order.schema.js
@@ -118,7 +118,7 @@ type OrderData {
     delivery_info: CustomerAddressType
     delivery_time: String
     customer_discount_amount: Float
-    vat: Int
+    vat: Float
     payment_type: String
     createdAt: String
     updatedAt: String
